Format move

+++ Galaxy_Wizard/Bishop.h
@@ -4,7 +4,7 @@ class Bishop :
 	public Figure
 {
 public:
-	Bishop(size_t x, size_t y, WORD Score);
+	Bishop(size_t x, size_t y, DWORD Score);
 	~Bishop();
 
 	virtual Figure* Clone();
