+++ Galaxy_Wizard/Figure.cpp
@@ -2,7 +2,7 @@
 #include "Figure.h"
 
 
-Figure::Figure(size_t x, size_t y, WORD Score)
+Figure::Figure(size_t x, size_t y, DWORD Score)
 	:m(x), n(y), Value(Score)
 {
 }
