+++ Galaxy_Wizard/Pawn.h
@@ -4,7 +4,7 @@ class Pawn :
 	public Figure
 {
 public:
-	Pawn(size_t x, size_t y, WORD Score);
+	Pawn(size_t x, size_t y, DWORD Score);
 	~Pawn();
 
 	virtual Figure* Clone();
